refactor: ユーザ確認のメッセージのみ、クラスへ分離する。
