+++ src/components/main/tree/FolderTreeFactory.ts
@@ -1,7 +1,7 @@
 import TreeItem from "@/presentation/tree/TreeItem";
 import DiagramType from "@/domain/diagram/DiagramType";
-import Diagrams from "~/domain/diagram/Diagrams";
-import Diagram from "~/domain/diagram/Diagram";
+import Diagrams from "@/domain/diagram/Diagrams";
+import Diagram from "@/domain/diagram/Diagram";
 import Folder from "./Folder";
 
 export default class FolderTreeFactory {
