+++ src/components/main/tree/TreeWrapper.ts
@@ -1,6 +1,6 @@
 import TreeItem from "@/presentation/tree/TreeItem";
 import DiagramType from "@/domain/diagram/DiagramType";
-import Diagram from "~/domain/diagram/Diagram";
+import Diagram from "@/domain/diagram/Diagram";
 import Folder from "./Folder";
 import FolderTreeFactory from "./FolderTreeFactory";
